# postman/swagger.yaml
openapi: 3.0.0
info:
  title: Subscriber API
  description: |-
    List, Add, Update and Delete Subscribers
  termsOfService: http://swagger.io/terms/
  contact: 
    name: CompTia
    url: https://www.comptia.com/
    email: dng-root@unnst.com
  license:
    name: MIT
    url: https://spdx.org/licenses/MIT.html
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Development server
paths:   
  /api/items:
    get:
      summary: Retrieve a list of subscribers
      responses:
        200:
          description: A list of subscribers
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  subscribedToChannel:
                    type: string
 

  /api/item/:
    post:
      summary: Create a new subscriber
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: John Doe
                subscribedToChannel:
                  type: string
                  example: Channel1
      responses:
        201:
          description: Subscriber created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  subscribedToChannel:
                    type: string


   
  /api/items/{id}:
    patch:
      summary: Update an existing subscriber
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The subscriber ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Jane Doe
                subscribedToChannel:
                  type: string
                  example: Channel2
      responses:
        200:
          description: Subscriber updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  subscribedToChannel:
                    type: string



   
  /api/item/{id}:
    delete:
      summary: Delete a subscriber
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: The subscriber ID
      responses:
        204:
          description: Subscriber deleted successfully